feat(rail): 分组栏、拖放提示与整组操作，修复五个阻断性缺陷
完成分组的完整交互闭环：拖放合并、分组栏显示、整组移动与缩放、
上下文菜单、关闭整组。补充三层集成测试。

最关键的一个缺陷是 System.Text.Json 无法序列化 nint。SessionMember.Handle
是 nint，导致每次分组时写会话快照都抛 NotSupportedException；而写快照排在
执行窗口指令和创建分组栏之前，异常一抛后面全不执行。更糟的是这个异常被
UiDispatcher 吞进了 Debug.WriteLine，发布运行时完全不可见。

表象是"组建出来了（后续拖拽报已在另一个组中）、窗口没对齐、分组栏没出现、
日志一片正常"，前后绕了好几轮才定位到。句柄改存 long，PersistSession 改为
捕获所有异常并继续（快照写不了是降级，因此完全不让用户分组是更糟的结果），
并把编排层和窗口过程的异常全部落盘。

其余四个：

SetWindowPos 的 hWndInsertAfter 语义是"排在该窗口之后"，我写成
SetWindowPos(轨道, 成员窗口, ...) 等于要求把轨道放到成员窗口后面，
正好是想要效果的反面，分组栏被完整盖住。改为移动成员窗口。

分组栏原本放在承载窗口上方，窗口贴着屏幕顶部时轨道被挤到屏幕外。
改为从窗口顶部预留轨道高度，内容区下移，与 Groupy 一致。

MoveGroup 和 OnLocationChanged 算出的新矩形只用在局部变量上，从未写回
组模型，下一帧仍从旧矩形算起，窗口被反复拉回原处，表现为"拖不动"。

菜单按钮弹出时传的目标是空句柄，导致针对具体标签的项全部灰显。
改为默认作用于当前活动标签。

行为变更：分组后整组是一个整体，拖任意成员则整组跟随移动，不再自动拆分。
拆分只能通过菜单显式发起。早期的"拖走即拆分"会让用户只想挪个位置就把组拆了，
而且拖回来不会自动合并。

视觉：撤销了修改成员窗口圆角偏好的做法 —— 那是在改用户应用的外观。
改为分组栏向下覆盖一段（等于窗口圆角半径）压住圆角区域消除接缝，
窗口本身分毫未动。方角窗口不额外覆盖。

拖动卡顿的三个来源：每次对齐调一次 DWM 属性设置（随上条移除）、
每帧每窗口查一次 DWM 边框（改为缓存，边框厚度只随样式与 DPI 变）、
指令堆积（改为同一时刻只允许一批对齐在途，丢弃过期中间帧）。

新增 --grouptest 走生产代码路径验证编排层，--railtest 验证分组栏可见性与
Z 序，--watch 实时观察拖放判定链路。三层缺一不可：曾出现 selftest 与
railtest 双绿但产品完全不可用，问题正藏在两者之间无人覆盖的编排代码里。

测试 148 个单元测试 + 三个集成测试全部通过。

+++ src/TabNest.Core/Persistence/SessionSnapshot.cs
@@ -5,14 +5,22 @@ namespace TabNest.Core.Persistence;
 /// <summary>一个成员窗口在快照中的记录。</summary>
 public sealed record SessionMember
 {
-    public required nint Handle { get; init; }
+    /// <summary>
+    /// 窗口句柄，存为 <see cref="long"/> 而不是 <see cref="nint"/>。
+    ///
+    /// System.Text.Json 明确不支持序列化 IntPtr，用 nint 会在写快照时抛
+    /// NotSupportedException。而写快照发生在执行窗口指令之前，异常一抛后面全不执行 ——
+    /// 表现为"组建出来了，但窗口没对齐、分组栏也没出现"，且异常还被上层吞掉，完全静默。
+    /// </summary>
+    public required long Handle { get; init; }
+
     public required int ProcessId { get; init; }
     public required long ProcessStartTicks { get; init; }
     public required string Title { get; init; }
     public required string ProcessName { get; init; }
     public required WindowSnapshot Snapshot { get; init; }
 
-    public WindowIdentity ToIdentity() => new(Handle, ProcessId, ProcessStartTicks);
+    public WindowIdentity ToIdentity() => new((nint)Handle, ProcessId, ProcessStartTicks);
 }
 
 /// <summary>一个组在快照中的记录。</summary>
@@ -70,6 +78,7 @@ public sealed record SessionSnapshot
                 members.Add(new SessionMember
                 {
                     Handle = tab.Identity.Handle,
+
                     ProcessId = tab.Identity.ProcessId,
                     ProcessStartTicks = tab.Identity.ProcessStartTicks,
                     Title = tab.Title,
